requirements.txt
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
streamlit==1.29.0
pandas==2.1.4
beautifulsoup4==4.12.2
python-dotenv==1.0.0
schedule==1.2.0
SQLAlchemy==2.0.15
plotly==5.18.0
psycopg2-binary==2.9.9
google-auth-httplib2==0.1.1
